Always and unconditionally store all login credentials encrypted
While the way the credentials are encrypted is a bit bogus, it's still safer
than storing them in plain text, at any time. So, remove the pyDes import
error handling and let it fail hard instead, and remove all conditionals and
the Kodi setting that control if encryption takes place or not. Having
(script.module.)pyDes is a *requirement* anyway.