Replace script.module.pydes with script.module.pycryptodome
3DES encryption/decryption (currently only used for storing the Sky Go login
password) can aswell be done by functionality provided by pycryptodome, which
is shipped with Kodi as of now, so there's no need to depend on a module
(which is script.module.pydes) which has to be installed from a third party
repository (or, manually even). So, use/include/import pycryptodome instead
of pyDes and replace all 3DES handling with functionality provided by
pycryptodome.